Refactor code to use yargs and restructure utility functions
The commit includes the change of the earlier function-based approach to an Object-Oriented way for handling WordPress package in the index. This refactor was made to increase readability and modularity of the code.

The utility functions have been merged into a single `utils.js` file, which was moved to 'lib' directory and reexported as an object for simpler imports. Also, to increase readability, 'getOptionValue' was replaced with 'yargs' and 'hideBin' for more efficient and streamlined command-line parameters parsing.

The revision also includes addition of 'getConfig' function to handle reading and parsing of configuration file (wp-package.json).

index.js was also renamed and relocated to 'lib' folder for better project structure and increased code maintainability.

A new dependency to yargs library was added to parse command line parameters.

The handcrafted function to parse command line arguments was dropped in favour of popular 'yargs' library for increased reliability and easier maintenance.

The binary and main entry point in the package.json was updated to reflect the new file location. The version was also bumped.

The package-lock.json was automatically updated to adjust the new 'yargs' dependency and version bump.