Add WordPress configuration dump tool
Added new file `dump.js` for creating and saving a JSON dump of a WordPress configuration. This includes information about the WordPress version, installed themes and plugins. The dumped configuration is saved in a `wp-package.json` file in the root directory. This feature enables the user to easily retrieve and check the details of their WordPress configuration.