Add additional usage instructions and details to readme.md
This commit enhances the 'Usage' section of 'readme.md' by adding instructions for creating a configuration file using the `npx wpmm --init` and `npx wpmm --dump` commands. Furthermore, it also introduces a new 'Commands and Options' section. This section explains the three main commands and their respective usage. There are also modifications to the 'Configuration' section with added options and the removal of some redundancies. Lastly, 'Version Detection' was added to the 'Features' list. This commit aims to make the project's usage and configuration more straightforward and intuitive to users.