Refactor code and enhance database functionality
Refactored the code by moving functions to newly created utils/index.js, improving the organization of utility functions. Added `getInfo` function to the utils import. Renamed WordPressInstaller to Installer, WordPressConfigDump to Dump, and added Database class for handling new database operations. Furthermore, the handling of the 'dump' and 'init' actions have been updated along with the addition of actions for 'upload-database' and 'dump-database'.