Refactor getConfig function and connection settings in data.js
The getConfig function in 'data.js' file has been refactored to handle scenarios when a template is not provided and add an "initConfig" function for initialising the configuration with the base folder. The parameters in the getConnectionSettings function have also been modified to align with the new configuration settings format.