Replace axios.get with axiosFetch in getLastWp function
The getLastWp function in wordpress.js has been updated to use the axiosFetch function instead of axios.get. In addition to the replacement, error handling has been improved: now in case of an error or if no offers are received from the WordPress API, an appropriate message is logged to the console instead of throwing an exception.